backend/microbiome-backend/requirements-prod.txt
# Production requirements - add these to requirements.txt
boto3>=1.34.0
django-storages[s3]>=1.14.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
